# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TrackStar API",
    "description": "Personal Task & Habit Tracker API",
    "version": "1.0.0",
    "contact": {
      "name": "TrackStar Team",
      "email": "support@trackstar.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/habits/": {
      "get": {
        "tags": [
          "Habits"
        ],
        "summary": "Get all habits for the current user",
        "description": "Retrieve all habits belonging to the authenticated user.",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Habits"
        ],
        "summary": "Create a new habit",
        "description": "Create a new habit for the authenticated user.",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Habit information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Daily Exercise"
                },
                "description": {
                  "type": "string",
                  "example": "Do 30 minutes of exercise daily"
                },
                "frequency": {
                  "type": "string",
                  "example": "daily"
                },
                "targetCount": {
                  "type": "number",
                  "example": 1
                },
                "category": {
                  "type": "string",
                  "example": "health"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/habits/{id}": {
      "get": {
        "tags": [
          "Habits"
        ],
        "summary": "Get a specific habit by ID",
        "description": "Retrieve details of a specific habit by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Habit ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Habits"
        ],
        "summary": "Update an existing habit",
        "description": "Update details of an existing habit by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Habit ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated habit information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Updated habit name"
                },
                "description": {
                  "type": "string",
                  "example": "Updated description"
                },
                "frequency": {
                  "type": "string",
                  "example": "weekly"
                },
                "isActive": {
                  "type": "boolean",
                  "example": false
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Habits"
        ],
        "summary": "Delete a habit",
        "description": "Delete a specific habit by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Habit ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/habit-logs/": {
      "get": {
        "tags": [
          "Habit Logs"
        ],
        "summary": "Get all habit logs for current user",
        "description": "Retrieve all habit completion logs for the authenticated user with optional filtering.",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "habitId",
            "in": "query",
            "description": "Filter by specific habit ID",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Filter logs from this date",
            "required": false,
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Filter logs until this date",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of logs to return",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Habit Logs"
        ],
        "summary": "Log a habit completion",
        "description": "Record a completion of a habit for the current user.",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Habit completion information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "habitId": {
                  "type": "string",
                  "example": "64a7b8c9d1234567890abcde"
                },
                "completedDate": {
                  "type": "string",
                  "example": "2024-01-15"
                },
                "completionCount": {
                  "type": "number",
                  "example": 1
                },
                "notes": {
                  "type": "string",
                  "example": "Completed morning workout"
                },
                "mood": {
                  "type": "string",
                  "example": "good"
                },
                "difficulty": {
                  "type": "number",
                  "example": 3
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/habit-logs/{id}": {
      "get": {
        "tags": [
          "Habit Logs"
        ],
        "summary": "Get a specific habit log by ID",
        "description": "Retrieve details of a specific habit completion log by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Habit Log ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Habit Logs"
        ],
        "summary": "Update a habit log",
        "description": "Update details of an existing habit completion log.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Habit Log ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated habit log information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "completionCount": {
                  "type": "number",
                  "example": 2
                },
                "notes": {
                  "type": "string",
                  "example": "Updated notes"
                },
                "mood": {
                  "type": "string",
                  "example": "excellent"
                },
                "difficulty": {
                  "type": "number",
                  "example": 2
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Habit Logs"
        ],
        "summary": "Delete a habit log",
        "description": "Delete a specific habit completion log by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Habit Log ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/habit-logs/stats/{habitId}": {
      "get": {
        "tags": [
          "Habit Logs"
        ],
        "summary": "Get statistics for a specific habit",
        "description": "Retrieve comprehensive statistics and analytics for a specific habit including completion rates, streaks, and performance metrics.",
        "parameters": [
          {
            "name": "habitId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Habit ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to analyze",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "User ID for authentication (temporary for testing)"
      }
    }
  }
}